+++ chalice/package.py
@@ -170,6 +170,11 @@ class SAMTemplateGenerator(object):
                 'DefinitionBody': resource.swagger_doc,
             }
         }
+        if resource.minimum_compression:
+            properties = resources['RestAPI']['Properties']
+            properties['MinimumCompressionSize'] = \
+                int(resource.minimum_compression)
+
         handler_cfn_name = to_cfn_resource_name(
             resource.lambda_function.resource_name)
         api_handler = template['Resources'].pop(handler_cfn_name)