+++ tests/unit/test_package.py
@@ -257,10 +257,23 @@ class TestSAMTemplate(object):
             },
         }
 
+    def test_can_generate_rest_api_without_compression(
+            self, sample_app_with_auth):
+        config = Config.create(chalice_app=sample_app_with_auth,
+                               project_dir='.',
+                               api_gateway_stage='api',
+                               )
+        template = self.generate_template(config, 'dev')
+        resources = template['Resources']
+        assert 'MinimumCompressionSize' not in \
+            resources['RestAPI']['Properties']
+
     def test_can_generate_rest_api(self, sample_app_with_auth):
         config = Config.create(chalice_app=sample_app_with_auth,
                                project_dir='.',
-                               api_gateway_stage='api')
+                               api_gateway_stage='api',
+                               minimum_compression_size=100,
+                               )
         template = self.generate_template(config, 'dev')
         resources = template['Resources']
         # Lambda function should be created.
@@ -279,6 +292,8 @@ class TestSAMTemplate(object):
                         {'RestAPIId': {'Ref': 'RestAPI'}}]}},
         }
         assert resources['RestAPI']['Type'] == 'AWS::Serverless::Api'
+        assert resources['RestAPI']['Properties']['MinimumCompressionSize'] \
+            == 100
         # We should also create the auth lambda function.
         assert resources['Myauth']['Type'] == 'AWS::Serverless::Function'
         # Along with permission to invoke from API Gateway.